pyproject.toml
[tool.poetry]
name = "bobsled"
version = "2.0.0"
description = ""
authors = ["James Turk <dev@jamesturk.net>"]

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^19.1"
pytest = "^5.0"
pyyaml = "^5.1"
docker = "^4.0"
starlette = "^0.12.4"
uvicorn = "^0.8.4"
aiofiles = "^0.4.0"
jinja2 = "^2.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
